refactor: extract ZoteroLibrary to legacy_sources.py to fix file size violation
- Extract legacy ZoteroLibrary class (142 lines) to new file legacy_sources.py
- Update all imports to reference new location
- Reduce sources.py from 1068 lines to 925 lines (under 1000 line limit)
- All tests pass with no functionality changes
