+++ puby/__init__.py
@@ -4,9 +4,10 @@ __version__ = "0.1.0"
 
 from .client import PublicationClient
 from .env import get_api_key, load_api_keys
+from .legacy_sources import ZoteroLibrary
 from .matcher import PublicationMatcher
 from .models import Author, Publication
-from .sources import ORCIDSource, PureSource, ScholarSource, ZoteroLibrary
+from .sources import ORCIDSource, PureSource, ScholarSource
 
 __all__ = [
     "Author",
